Change KeySize parameter from byte to int
- Changed parameter keySize in GetSlicedKey from byte to key to ensure future compatibility as const PasswordHelper.KeySize is saved as an int

+++ VaultCrypt.Tests/FakeVaultSession.cs
@@ -68,7 +68,7 @@ namespace VaultCrypt.Tests
 
         public void RaiseEncryptedFileListUpdated() => RaiseEncryptedFileListUpdatedWasCalled = true;
 
-        public ReadOnlySpan<byte> GetSlicedKey(byte keySize)
+        public ReadOnlySpan<byte> GetSlicedKey(int keySize)
         {
             GetSlicedKeyWasCalled = true;
             return new byte[0];
